requirements.txt
Flask
google-generativeai
speechrecognition
pyttsx3
pyaudio
pyautogui
